Tidied buffers.

+++ cubes/gl/include/egl.h
@@ -29,7 +29,7 @@ private:
     GLuint EBO;
     Grid grid;
     std::vector<Vertex> vertices;
-    std::vector<uint16_t> indices;
+    std::vector<unsigned int> indices;
 
     void initWindow();
     void initGL();
